chore(mark): remove importing of scanner report data
In RAVE, the voter station will not need to import data from any scanners.
