+++ apps/mark/backend/src/app.test.ts
@@ -1,4 +1,5 @@
-import { assert, err, ok, Optional, Result } from '@votingworks/basics';
+import { InsertedSmartCardAuthApi } from '@votingworks/auth';
+import { assert } from '@votingworks/basics';
 import {
   electionFamousNames2021Fixtures,
   electionSampleDefinition,
@@ -6,31 +7,22 @@ import {
 } from '@votingworks/fixtures';
 import {
   fakeElectionManagerUser,
-  fakePollWorkerUser,
   fakeSessionExpiresAt,
   mockOf,
   suppressingConsoleOutput,
 } from '@votingworks/test-utils';
-import { InsertedSmartCardAuthApi } from '@votingworks/auth';
-import {
-  ALL_PRECINCTS_SELECTION,
-  ReportSourceMachineType,
-  safeParseSystemSettings,
-  ScannerReportData,
-  ScannerReportDataSchema,
-} from '@votingworks/utils';
+import { safeParseSystemSettings } from '@votingworks/utils';
 
-import { Buffer } from 'buffer';
 import { createBallotPackageZipArchive, MockUsb } from '@votingworks/backend';
-import { Server } from 'http';
 import * as grout from '@votingworks/grout';
 import {
   DEFAULT_SYSTEM_SETTINGS,
   safeParseJson,
   SystemSettingsSchema,
-  TEST_JURISDICTION,
 } from '@votingworks/types';
-import { configureApp, createApp } from '../test/app_helpers';
+import { Buffer } from 'buffer';
+import { Server } from 'http';
+import { createApp } from '../test/app_helpers';
 import { Api } from './app';
 
 let apiClient: grout.Client<Api>;
@@ -45,8 +37,6 @@ beforeEach(() => {
 afterEach(() => {
   server?.close();
 });
-
-const jurisdiction = TEST_JURISDICTION;
 
 test('uses machine config from env', async () => {
   const originalEnv = process.env;
@@ -71,106 +61,6 @@ test('uses default machine config if not set', async () => {
     machineId: '0000',
     codeVersion: 'dev',
     screenOrientation: 'portrait',
-  });
-});
-
-test('read scanner report data from card', async () => {
-  const { electionDefinition } = electionFamousNames2021Fixtures;
-  const { electionHash } = electionDefinition;
-  await configureApp(apiClient, mockAuth, mockUsb);
-
-  const scannerReportData: ScannerReportData = {
-    ballotCounts: {},
-    isLiveMode: false,
-    machineId: '0000',
-    pollsTransition: 'close_polls',
-    precinctSelection: ALL_PRECINCTS_SELECTION,
-    tally: [],
-    tallyMachineType: ReportSourceMachineType.PRECINCT_SCANNER,
-    timePollsTransitioned: 0,
-    timeSaved: 0,
-    totalBallotsScanned: 0,
-  };
-  mockOf(mockAuth.readCardData).mockImplementation(() =>
-    Promise.resolve(ok(scannerReportData))
-  );
-
-  let result: Result<Optional<ScannerReportData>, Error>;
-
-  mockOf(mockAuth.getAuthStatus).mockImplementation(() =>
-    Promise.resolve({ status: 'logged_out', reason: 'no_card' })
-  );
-  result = await apiClient.readScannerReportDataFromCard();
-  expect(result).toEqual(err(new Error('User is not logged in')));
-
-  mockOf(mockAuth.getAuthStatus).mockImplementation(() =>
-    Promise.resolve({
-      status: 'logged_in',
-      user: fakeElectionManagerUser(electionDefinition),
-      sessionExpiresAt: fakeSessionExpiresAt(),
-    })
-  );
-  result = await apiClient.readScannerReportDataFromCard();
-  expect(result).toEqual(err(new Error('User is not a poll worker')));
-
-  mockOf(mockAuth.getAuthStatus).mockImplementation(() =>
-    Promise.resolve({
-      status: 'logged_in',
-      user: fakePollWorkerUser(electionDefinition),
-      sessionExpiresAt: fakeSessionExpiresAt(),
-    })
-  );
-  result = await apiClient.readScannerReportDataFromCard();
-  expect(result).toEqual(ok(scannerReportData));
-  expect(mockAuth.readCardData).toHaveBeenCalledTimes(1);
-  expect(mockAuth.readCardData).toHaveBeenNthCalledWith(
-    1,
-    { ...DEFAULT_SYSTEM_SETTINGS, electionHash, jurisdiction },
-    { schema: ScannerReportDataSchema }
-  );
-});
-
-test('clear scanner report data from card', async () => {
-  const { electionDefinition } = electionFamousNames2021Fixtures;
-  const { electionHash } = electionDefinition;
-  await configureApp(apiClient, mockAuth, mockUsb);
-
-  mockOf(mockAuth.clearCardData).mockImplementation(() =>
-    Promise.resolve(ok())
-  );
-
-  let result: Result<void, Error>;
-
-  mockOf(mockAuth.getAuthStatus).mockImplementation(() =>
-    Promise.resolve({ status: 'logged_out', reason: 'no_card' })
-  );
-  result = await apiClient.clearScannerReportDataFromCard();
-  expect(result).toEqual(err(new Error('User is not logged in')));
-
-  mockOf(mockAuth.getAuthStatus).mockImplementation(() =>
-    Promise.resolve({
-      status: 'logged_in',
-      user: fakeElectionManagerUser(electionDefinition),
-      sessionExpiresAt: fakeSessionExpiresAt(),
-    })
-  );
-  result = await apiClient.clearScannerReportDataFromCard();
-  expect(result).toEqual(err(new Error('User is not a poll worker')));
-
-  mockOf(mockAuth.getAuthStatus).mockImplementation(() =>
-    Promise.resolve({
-      status: 'logged_in',
-      user: fakePollWorkerUser(electionDefinition),
-      sessionExpiresAt: fakeSessionExpiresAt(),
-    })
-  );
-  result = await apiClient.clearScannerReportDataFromCard();
-  expect(result).toEqual(ok());
-  expect(mockAuth.clearCardData).toHaveBeenCalledTimes(1);
-  expect(mockAuth.clearCardData).toHaveBeenNthCalledWith(1, {
-    ...DEFAULT_SYSTEM_SETTINGS,
-    electionHash,
-    jurisdiction,
   });
 });
