+++ apps/mark/frontend/src/app_card_unhappy_paths.test.tsx
@@ -1,8 +1,5 @@
 import React from 'react';
-import { electionSampleDefinition } from '@votingworks/fixtures';
-import { hasTextAcrossElements } from '@votingworks/test-utils';
 import { MemoryStorage, MemoryHardware } from '@votingworks/utils';
-import { err } from '@votingworks/basics';
 import { render, screen } from '../test/react_testing_library';
 
 import { App } from './app';
@@ -26,44 +23,6 @@ beforeEach(() => {
 
 afterEach(() => {
   apiMock.mockApiClient.assertComplete();
-});
-
-test('Poll worker card with invalid scanner report data is treated like card without scanner report data', async () => {
-  // ====================== BEGIN CONTEST SETUP ====================== //
-
-  const hardware = MemoryHardware.buildStandard();
-  const storage = new MemoryStorage();
-  apiMock.expectGetMachineConfig();
-
-  await setElectionInStorage(storage, electionSampleDefinition);
-  await setStateInStorage(storage, {
-    pollsState: 'polls_closed_initial',
-  });
-
-  render(
-    <App
-      hardware={hardware}
-      storage={storage}
-      apiClient={apiMock.mockApiClient}
-      reload={jest.fn()}
-    />
-  );
-  await advanceTimersAndPromises();
-
-  // ====================== END CONTEST SETUP ====================== //
-
-  screen.getByText('Insert Poll Worker card to open.');
-
-  apiMock.setAuthStatusPollWorkerLoggedIn(electionSampleDefinition, {
-    scannerReportDataReadResult: err(new Error('Invalid scanner report data')),
-  });
-  await advanceTimersAndPromises();
-
-  // Land on pollworker screen
-  await screen.findByText(hasTextAcrossElements('Polls: Closed'));
-
-  // No prompt to print precinct tally report
-  expect(screen.queryAllByText('Tally Report on Card')).toHaveLength(0);
 });
 
 test('Shows card backwards screen when card connection error occurs', async () => {
