+++ apps/mark/frontend/src/app_cardless_voting.test.tsx
@@ -107,8 +107,7 @@ test('Cardless Voting Flow', async () => {
   // Open Polls with Poll Worker Card
   apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition);
   await advanceTimersAndPromises();
-  userEvent.click(await screen.findByText('Open Polls'));
-  fireEvent.click(screen.getByText('Open Polls on VxMark Now'));
+  userEvent.click(await screen.findButton('Open Polls'));
 
   // Remove card
   apiMock.setAuthStatusLoggedOut();
@@ -126,7 +125,6 @@ test('Cardless Voting Flow', async () => {
     .resolves();
   fireEvent.click(within(screen.getByTestId('ballot-styles')).getByText('12'));
   apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {
-    isScannerReportDataReadExpected: false,
     cardlessVoterUserParams: {
       ballotStyleId: '12',
       precinctId: '23',
@@ -137,9 +135,7 @@ test('Cardless Voting Flow', async () => {
   // Poll Worker deactivates ballot style
   apiMock.mockApiClient.endCardlessVoterSession.expectCallWith().resolves();
   userEvent.click(await screen.findByText('Deactivate Voting Session'));
-  apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {
-    isScannerReportDataReadExpected: false,
-  });
+  apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {});
   await screen.findByText('Select Voter’s Ballot Style');
 
   // Poll Worker reactivates ballot style
@@ -148,7 +144,6 @@ test('Cardless Voting Flow', async () => {
     .resolves();
   fireEvent.click(within(screen.getByTestId('ballot-styles')).getByText('12'));
   apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {
-    isScannerReportDataReadExpected: false,
     cardlessVoterUserParams: {
       ballotStyleId: '12',
       precinctId: '23',
@@ -184,9 +179,7 @@ test('Cardless Voting Flow', async () => {
   // Poll Worker resets ballot to remove votes
   apiMock.mockApiClient.endCardlessVoterSession.expectCallWith().resolves();
   fireEvent.click(screen.getByText('Reset Ballot'));
-  apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {
-    isScannerReportDataReadExpected: false,
-  });
+  apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {});
 
   // Back on Poll Worker screen
   await screen.findByText('Select Voter’s Ballot Style');
@@ -197,7 +190,6 @@ test('Cardless Voting Flow', async () => {
     .resolves();
   fireEvent.click(within(screen.getByTestId('ballot-styles')).getByText('12'));
   apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {
-    isScannerReportDataReadExpected: false,
     cardlessVoterUserParams: {
       ballotStyleId: '12',
       precinctId: '23',
@@ -293,7 +285,6 @@ test('Another Voter submits blank ballot and clicks Done', async () => {
     within(await screen.findByTestId('ballot-styles')).getByText('12')
   );
   apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {
-    isScannerReportDataReadExpected: false,
     cardlessVoterUserParams: {
       ballotStyleId: '12',
       precinctId: '23',
@@ -410,8 +401,7 @@ test('poll worker must select a precinct first', async () => {
   // Open Polls with Poll Worker Card
   apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition);
   await advanceTimersAndPromises();
-  userEvent.click(await screen.findByText('Open Polls'));
-  fireEvent.click(screen.getByText('Open Polls on VxMark Now'));
+  userEvent.click(await screen.findButton('Open Polls'));
 
   // Remove card
   apiMock.setAuthStatusLoggedOut();
@@ -433,7 +423,6 @@ test('poll worker must select a precinct first', async () => {
     .resolves();
   fireEvent.click(within(screen.getByTestId('ballot-styles')).getByText('12'));
   apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {
-    isScannerReportDataReadExpected: false,
     cardlessVoterUserParams: {
       ballotStyleId: '12',
       precinctId: '23',
@@ -444,9 +433,7 @@ test('poll worker must select a precinct first', async () => {
   // Poll Worker deactivates ballot style
   apiMock.mockApiClient.endCardlessVoterSession.expectCallWith().resolves();
   fireEvent.click(screen.getByText('Deactivate Voting Session'));
-  apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {
-    isScannerReportDataReadExpected: false,
-  });
+  apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {});
   await screen.findByText('2. Select Voter’s Ballot Style');
 
   // Poll Worker reactivates ballot style
@@ -458,7 +445,6 @@ test('poll worker must select a precinct first', async () => {
     .resolves();
   fireEvent.click(within(screen.getByTestId('ballot-styles')).getByText('12'));
   apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {
-    isScannerReportDataReadExpected: false,
     cardlessVoterUserParams: {
       ballotStyleId: '12',
       precinctId: '23',
@@ -494,9 +480,7 @@ test('poll worker must select a precinct first', async () => {
   // Poll Worker resets ballot to remove votes
   apiMock.mockApiClient.endCardlessVoterSession.expectCallWith().resolves();
   fireEvent.click(screen.getByText('Reset Ballot'));
-  apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {
-    isScannerReportDataReadExpected: false,
-  });
+  apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {});
 
   // Back on Poll Worker screen
   await screen.findByText('1. Select Voter’s Precinct');
@@ -511,7 +495,6 @@ test('poll worker must select a precinct first', async () => {
     .resolves();
   fireEvent.click(within(screen.getByTestId('ballot-styles')).getByText('12'));
   apiMock.setAuthStatusPollWorkerLoggedIn(electionDefinition, {
-    isScannerReportDataReadExpected: false,
     cardlessVoterUserParams: {
       ballotStyleId: '12',
       precinctId: '23',
