+++ apps/mark/frontend/src/pages/poll_worker_screen.test.tsx
@@ -14,8 +14,6 @@ import {
 } from '@votingworks/test-utils';
 import userEvent from '@testing-library/user-event';
 
-import { Logger, LogSource } from '@votingworks/logging';
-import { ok } from '@votingworks/basics';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { fireEvent, screen } from '../../test/react_testing_library';
 
@@ -81,7 +79,6 @@ function renderScreen(
           screenReader={new AriaScreenReader(fakeTts())}
           updatePollsState={jest.fn()}
           reload={jest.fn()}
-          logger={new Logger(LogSource.VxMarkFrontend)}
           {...props}
         />
       </QueryClientProvider>
@@ -89,21 +86,7 @@ function renderScreen(
   );
 }
 
-test('renders PollWorkerScreen in MarkAndPrint app mode', () => {
-  apiMock.mockApiClient.readScannerReportDataFromCard
-    .expectCallWith()
-    .resolves(ok(undefined));
-  renderScreen(undefined, undefined, undefined);
-  screen.getByText('Poll Worker Actions');
-  expect(
-    screen.getByText('Ballots Printed:').parentElement!.textContent
-  ).toEqual('Ballots Printed: 0');
-});
-
 test('renders PollWorkerScreen', () => {
-  apiMock.mockApiClient.readScannerReportDataFromCard
-    .expectCallWith()
-    .resolves(ok(undefined));
   renderScreen();
   screen.getByText('Poll Worker Actions');
 });
@@ -113,9 +96,6 @@ test('switching out of test mode on election day', () => {
     ...electionSampleWithSeal,
     date: new Date().toISOString(),
   });
-  apiMock.mockApiClient.readScannerReportDataFromCard
-    .expectCallWith()
-    .resolves(ok(undefined));
   const enableLiveMode = jest.fn();
   renderScreen({
     pollWorkerAuth: fakePollWorkerAuth(electionDefinition),
@@ -135,9 +115,6 @@ test('keeping test mode on election day', () => {
     ...electionSampleWithSeal,
     date: new Date().toISOString(),
   });
-  apiMock.mockApiClient.readScannerReportDataFromCard
-    .expectCallWith()
-    .resolves(ok(undefined));
   const enableLiveMode = jest.fn();
   renderScreen({ electionDefinition, enableLiveMode });
 
@@ -149,9 +126,6 @@ test('keeping test mode on election day', () => {
 });
 
 test('live mode on election day', () => {
-  apiMock.mockApiClient.readScannerReportDataFromCard
-    .expectCallWith()
-    .resolves(ok(undefined));
   renderScreen({ isLiveMode: true });
   expect(
     screen.queryByText(
@@ -161,9 +135,6 @@ test('live mode on election day', () => {
 });
 
 test('navigates to System Diagnostics screen', () => {
-  apiMock.mockApiClient.readScannerReportDataFromCard
-    .expectCallWith()
-    .resolves(ok(undefined));
   const { unmount } = renderScreen();
 
   userEvent.click(screen.getByRole('button', { name: 'View More Actions' }));
@@ -180,36 +151,7 @@ test('navigates to System Diagnostics screen', () => {
   unmount();
 });
 
-test('requires confirmation to open polls if no report on card', () => {
-  apiMock.mockApiClient.readScannerReportDataFromCard
-    .expectCallWith()
-    .resolves(ok(undefined));
-  const updatePollsState = jest.fn();
-  renderScreen({
-    pollsState: 'polls_closed_initial',
-    updatePollsState,
-  });
-
-  fireEvent.click(screen.getByText('Open Polls'));
-
-  // Should show the modal and not open/close polls
-  expect(updatePollsState).not.toHaveBeenCalled();
-  screen.getByText('No Polls Opened Report on Card');
-
-  // Clicking Cancel closes the modal
-  fireEvent.click(screen.getByText('Cancel'));
-  screen.getByText('Open Polls');
-
-  // Clicking Open Polls on VxMark Now should open/close polls anyway
-  fireEvent.click(screen.getByText('Open Polls'));
-  fireEvent.click(screen.getByText('Open Polls on VxMark Now'));
-  expect(updatePollsState).toHaveBeenCalled();
-});
-
 test('can toggle between vote activation and "other actions" during polls open', async () => {
-  apiMock.mockApiClient.readScannerReportDataFromCard
-    .expectCallWith()
-    .resolves(ok(undefined));
   renderScreen({
     pollsState: 'polls_open',
     machineConfig: fakeMachineConfig(),
