+++ apps/mark/frontend/test/helpers/mock_api_client.tsx
@@ -19,8 +19,7 @@ import {
   fakeSessionExpiresAt,
   fakeSystemAdministratorUser,
 } from '@votingworks/test-utils';
-import { err, ok, Optional, Result } from '@votingworks/basics';
-import { ScannerReportData } from '@votingworks/utils';
+import { err, ok, Result } from '@votingworks/basics';
 import { ApiClientContext, createQueryClient } from '../../src/api';
 import { fakeMachineConfig } from './fake_machine_config';
 
@@ -87,26 +86,11 @@ export function createApiMock() {
     setAuthStatusPollWorkerLoggedIn(
       electionDefinition: ElectionDefinition,
       options: {
-        isScannerReportDataReadExpected?: boolean;
-        scannerReportDataReadResult?: Result<
-          Optional<ScannerReportData>,
-          Error
-        >;
         cardlessVoterUserParams?: CardlessVoterUserParams;
       } = {}
     ) {
       const { electionHash } = electionDefinition;
-      const {
-        isScannerReportDataReadExpected = true,
-        scannerReportDataReadResult = ok(undefined),
-        cardlessVoterUserParams,
-      } = options;
-
-      if (isScannerReportDataReadExpected) {
-        mockApiClient.readScannerReportDataFromCard
-          .expectCallWith()
-          .resolves(scannerReportDataReadResult);
-      }
+      const { cardlessVoterUserParams } = options;
 
       setAuthStatus({
         status: 'logged_in',
