+++ libs/utils/src/index.ts
@@ -3,7 +3,6 @@ export * from './as_boolean';
 export * from './auth';
 export * from './ballot_package';
 export * from './pins';
-export * from './compressed_tallies';
 export * from './date';
 export * from './elections';
 export * from './environment_variable';
